Перенесла вертску таблицы  из папки pages в featuers, обновила роут, временно перенесла мок данные и типы также из pages в features
